Move duplicate function to separate file
